convert base login component and ChooseAccountForm
